fix: work with request and response in controller favorite entity

+++ src/favorite/favorite.controller.ts
@@ -1,5 +1,7 @@
 import { Controller, Get, Post, Param, Delete, HttpCode } from '@nestjs/common';
+import { isValidId } from 'src/utils';
 import { FavoriteService } from './favorite.service';
+import { HttpException, HttpStatus } from '@nestjs/common';
 
 @Controller('favs')
 export class FavoriteController {
@@ -13,36 +15,105 @@ export class FavoriteController {
   @Post('track/:id')
   @HttpCode(201)
   addTrack(@Param('id') id: string) {
-    return this.favoriteService.addTrack(id);
+    if (isValidId(id)) {
+      const result = this.favoriteService.addTrack(id);
+      if (result) {
+        return {
+          message:
+            'track with id === trackId exists, and was successfully added to favorites',
+        };
+      }
+      throw new HttpException(
+        "record with id === trackId doesn't exist",
+        HttpStatus.UNPROCESSABLE_ENTITY,
+      );
+    }
+    throw new HttpException('id is invalid (not uuid)', HttpStatus.BAD_REQUEST);
   }
 
   @Post('album/:id')
   @HttpCode(201)
   addAlbum(@Param('id') id: string) {
-    return this.favoriteService.addAlbum(id);
+    if (isValidId(id)) {
+      const result = this.favoriteService.addAlbum(id);
+      if (result) {
+        return {
+          message:
+            'track with id === albumId exists, and was successfully added to favorites',
+        };
+      }
+      throw new HttpException(
+        "record with id === albumId doesn't exist",
+        HttpStatus.UNPROCESSABLE_ENTITY,
+      );
+    }
+    throw new HttpException('id is invalid (not uuid)', HttpStatus.BAD_REQUEST);
   }
 
   @Post('artist/:id')
   @HttpCode(201)
   addArtist(@Param('id') id: string) {
-    return this.favoriteService.addArtist(id);
+    if (isValidId(id)) {
+      const result = this.favoriteService.addArtist(id);
+      if (result) {
+        return {
+          message:
+            'track with id === artistId exists, and was successfully added to favorites',
+        };
+      }
+      throw new HttpException(
+        "record with id === artistId doesn't exist",
+        HttpStatus.UNPROCESSABLE_ENTITY,
+      );
+    }
+    throw new HttpException('id is invalid (not uuid)', HttpStatus.BAD_REQUEST);
   }
 
   @Delete('album/:id')
   @HttpCode(204)
   removeAlbum(@Param('id') id: string) {
-    this.favoriteService.removeAlbum(id);
+    if (isValidId(id)) {
+      const result = this.favoriteService.removeAlbum(id);
+
+      if (result) return;
+
+      throw new HttpException(
+        `corresponding album is not favorite`,
+        HttpStatus.NOT_FOUND,
+      );
+    }
+    throw new HttpException('id is invalid (not uuid)', HttpStatus.BAD_REQUEST);
   }
 
   @Delete('track/:id')
   @HttpCode(204)
   removeTrack(@Param('id') id: string) {
-    this.favoriteService.removeTrack(id);
+    if (isValidId(id)) {
+      const result = this.favoriteService.removeTrack(id);
+
+      if (result) return;
+
+      throw new HttpException(
+        `corresponding track is not favorite`,
+        HttpStatus.NOT_FOUND,
+      );
+    }
+    throw new HttpException('id is invalid (not uuid)', HttpStatus.BAD_REQUEST);
   }
 
   @Delete('artist/:id')
   @HttpCode(204)
   removeArtist(@Param('id') id: string) {
-    this.favoriteService.removeArtist(id);
+    if (isValidId(id)) {
+      const result = this.favoriteService.removeArtist(id);
+
+      if (result) return;
+
+      throw new HttpException(
+        `corresponding artist is not favorite`,
+        HttpStatus.NOT_FOUND,
+      );
+    }
+    throw new HttpException('id is invalid (not uuid)', HttpStatus.BAD_REQUEST);
   }
 }
